mcp/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
google-generativeai
